GAS-198 Create connection tool
The connection tool uses the PMM inventory as a source for SSH and
database connections, providing a standardised way to connect regardless
of the technology. For scenarios where the user either wishes to save
typing arguments when calling the relevant tool, or where they need to
override something coming from the PMM inventory (such as needing to set
the SSH port, overriding an address, etc), a config can be used; the
default path is ~/.config/gascan/connect-py.json.

* Added `scripts/connect/connect.py` to provide connectivity based upon
  the PMM inventory; this tool uses netrc to provide credentials for PMM
  API requests and forces the use of tokens.
* Updated `main.go` to embed the tool, extract it during runtime and
  generate a short, sample config if there is not already one in place
* Updated unit tests to include test for extracting the tool
* Updated `sample-bundle` task to include the connection tool